fix: add `passive` listener to wheel event
The `wheel` event listener was blocking the main thread, which negatively impacting user experience, particularly on mobile devices.

By adding the `{ passive: true }` option to the `wheel` event listener, we indicate to the browser that the listener will not call `preventDefault()`. This allows the browser to perform default actions, such as scrolling, without waiting for the listener to finish, resulting in a smoother user experience.

Note: It silences only the warning, because there is no state in which the browser is able to scroll nativly.